// src/lib/siteConfig.d.ts
export declare const siteConfig: {
    siteName: string;
    siteUrl: string;
    supportEmail: string;
    whatsappNumber: string;
    whatsappMessage: string;
    facebookUrl: string;
    instagramUrl: string;
    googleAnalyticsId: any;
    facebookPixelId: any;
    digitalPrice: number;
    printPrice: number;
    originalPrice: number;
    currency: string;
    currencySymbol: string;
};
export declare function getWhatsAppLink(customMessage?: string): string;